requirements.txt

fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
langchain
langchain-community
langchain-openai
langchain-core
tiktoken
qdrant-client
sqlalchemy
asyncpg
psycopg2-binary
httpx
passlib==1.7.4
bcrypt==4.2.1
python-jose[cryptography]
python-multipart
PyJWT>=2.8.0
alembic
